refactor commandline application help
